Baloncuk boyutu bug fixlendi
